feat: add a nice log if it's already cleared

+++ index.js
@@ -5,7 +5,7 @@ const inquirer = require("inquirer");
 const file = require("./fileContent");
 const File = require("./File");
 const Folder = require("./Folder");
-const clearcraConfig = require('./clearcra.config');
+const clearcraConfig = require("./clearcra.config");
 
 /** Get files from config to be removed  */
 const filesToRemove = clearcraConfig.removeFiles.map(
@@ -78,6 +78,10 @@ const foldersToAdd = clearcraConfig.addFolders.map(
     .then((answers) => {
       // Remove files
       if (answers.action === actions.CLEAR_FILES) {
+        // Give a nice log if there is nothing to remove
+        if (filesToRemove.every((file) => file.isRemoved)) {
+          return console.log(chalk.gray("✨ Already cleared"));
+        }
         filesToRemove.forEach((file) => {
           file.remove();
         });
